Function advent_of_code_2021::day_4::play_bingo_until_last

fn play_bingo_until_last(
    numbers: &Vec<u8>, 
    cards: &Vec<BingoCard>
) -> (BingoCard, u8)

Iterate through the numbers, marking each card as appropriate. Very similar to play_bingo except it needs to keep going until all cards have won. This leads to some complexity to manage removing the cards from the iterator as we’re looping over the same list.