Struct advent_of_code_2021::util::grid::Grid

pub struct Grid {
    pub numbers: Vec<u8>,
    pub width: usize,
}

A representation of a 2D grid of u8s. Originally implemented for crate::day_9, another grid was needed for crate::day_11 and so common methods were extracted to this shared module

Fields
numbers: Vec<u8>

Store the numbers in a 1D list…

width: usize

…and use the width to determine the 1D offset as a 2D co-ordinate

Implementations

impl Grid

pub fn get_all_surrounds(&self, y: usize, x: usize) -> Vec<((usize, usize), u8)>

Iterate through the four orthogonal cells, collecting the 2 - 4 values into a vector. Include the co-ordinates in the returned vector so that Grid::get_basin can recursively expand the set of cells in the basin.

fn iterate_and_flash(&mut self) -> usize

This is the core logic to implement a single pass of the octopuses powering up required for both parts. It first iterates through all the cells incrementing them by one. Any that started at 9 are added to a queue of cells that have triggered a flash. We then take cells off the to flash list, adding them to a set of all co-ordinates that have flashed this iteration (so that we can 0 them later, and also ensure no cell flashes twice). If the cell was new to the flashes set, it also increments its neighbours, adding any that exceed 9 to the trigger queue. Once the queue has been exhausted, we iterate through the resulting set of co-ordinates that flashed this iteration, set them to 0 and return the size of the set, as this is the metric needed for both parts’ solutions.

fn count_flashes(&mut self, cycles: usize) -> usize

Solution to part one. Iterate the grid times, summing the flashes this causes.

fn run_until_sync(&mut self) -> usize

Solution to part two. Iterate the grid until the set of flashes is the same size as the grid, i.e. all cells triggered a flash. Return the number of iterations required to reach that point.

impl Grid

fn get_orthogonal_surrounds(
&self,
y: usize,
x: usize
) -> Vec<((usize, usize), u8)>

Iterate through the four orthogonal cells, collecting the 2 - 4 values into a vector. Include the co-ordinates in the returned vector so that Grid::get_basin can recursively expand the set of cells in the basin.

fn is_lowest(&self, y: usize, x: usize) -> bool

Is the provided grid cell a local minimum

fn get_low_points(&self) -> Vec<((usize, usize), u8)>

Return a list of the co-ordinates and values of all local minima

fn get_risk_level(&self) -> usize

The risk level of the grid is the sum of the risk level of each low point, which is the low point’s height plus one.

fn get_basin(&self, y: usize, x: usize) -> HashSet<(usize, usize)>

Recursively walk to higher points from a starting minimum, stopping at the watershed of height 9. Returns the set of co-ordinates found.

fn get_largest_basin_sizes(&self) -> Vec<usize>

Iterate through the local minima, find the basin size of each, and return the highest three sizes found

impl Grid

pub fn iter(&self) -> GridCoords<'_>

Helper to abstract iterating over the whole grid

pub fn get(&self, y: usize, x: usize) -> Option<u8>

Return the value at the given co-ordinates

pub fn set(&mut self, y: usize, x: usize, val: u8) -> bool

Update the value in a given cell

pub fn get_with_coords(&self, pos: usize) -> Option<((usize, usize), u8)>

Used by GridCoords::next and other iterators over the grid , e.g. Grid::iterate_and_flash to turn the current iterator position into the x/y co-ordinates and the value in that cell.

pub fn get_relative(
&self,
y: usize,
x: usize,
dy: isize,
dx: isize
) -> Option<((usize, usize), u8)>

Given a cell and a delta, return the new co-ordinates and the value at those co-ordinates if it is within the grid, None otherwise.

pub fn print(&self)

Dump the grid to stdout - useful for visualising the grid when debugging